+++ shiny/experimental/ui/_valuebox.py
@@ -9,8 +9,8 @@ from ...ui._x._htmldeps import value_box_dependency
 from ...ui.css_unit import CssUnit, as_css_unit, as_width_unit
 from ...ui.fill import as_fill_carrier
 from ._card import CardItem, card, card_body
-from ._layout import layout_column_wrap
-from ._utils import is_01_scalar
+from ...ui._layout import layout_column_wrap
+from ...ui._x._utils import is_01_scalar
 
 __all__ = (
     "value_box",
