+++ shiny/ui/_card.py
@@ -1,15 +1,27 @@
 from __future__ import annotations
 
-from typing import Optional
+from typing import Optional, Protocol
 
-from htmltools import Tag, TagAttrs, TagAttrValue, TagChild, TagFunction, TagList, tags
+from htmltools import (
+    HTML,
+    Tag,
+    TagAttrs,
+    TagAttrValue,
+    TagChild,
+    TagFunction,
+    TagList,
+    css,
+    div,
+    tags,
+)
 
 # from .._docstring import add_example
-from ..types import MISSING
-from .css_unit._css_unit import CssUnit
-
-from ..types import MISSING_TYPE, MISSING
+from ..types import MISSING, MISSING_TYPE
 from ._tag import consolidate_attrs
+from ._tooltip import tooltip
+from ._x._htmldeps import card_dependency
+from .css_unit._css_unit import CssUnit, as_css_padding, as_css_unit
+from .fill import as_fill_item, as_fillable_container
 
 # No fill / fillable!
 # TODO-barret; Handle `card_body()` in docs
@@ -110,8 +122,7 @@ def _card_impl(
     wrapper: WrapperCallable | None | MISSING_TYPE = MISSING,
     **kwargs: TagAttrValue,
 ) -> Tag:
-
-   if isinstance(wrapper, MISSING_TYPE):
+    if isinstance(wrapper, MISSING_TYPE):
         wrapper = card_body
 
     attrs, children = consolidate_attrs(*args, class_=class_, **kwargs)
@@ -244,7 +255,6 @@ def _wrap_children_in_card(
     card_items = _as_card_items(*children, wrapper=wrapper)
     tag_children = [card_item.resolve() for card_item in card_items]
     return tag_children
-
 
 
 # TODO-maindocs; @add_example()
